Project3/Solution/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
weaviate-client==4.4.0
langchain-openai==0.0.2
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
pandas==2.1.4
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
opencensus-ext-azure==1.1.13
asyncio==3.4.3
httpx==0.25.2